+++ client/ayon_houdini/plugins/publish/collect_arnold_rop.py
@@ -4,9 +4,7 @@ import re
 import hou
 import pyblish.api
 
-from ayon_core.pipeline import PublishError
 from ayon_houdini.api import plugin
-from ayon_houdini.api.lib import evalParmNoFrame
 
 
 class CollectArnoldROPRenderProducts(plugin.HoudiniInstancePlugin):
@@ -161,12 +159,3 @@ class CollectArnoldROPRenderProducts(plugin.HoudiniInstancePlugin):
                 os.path.join(dir, (file % i)).replace("\\", "/"))
 
         return expected_files
-
-    def evalParmNoFrame(self, rop, parm, **kwargs):
-        try:
-            return evalParmNoFrame(rop, parm, **kwargs)
-        except Exception as exc:
-            raise PublishError(
-                f"Failed evaluating parameter '{parm}' on Rop node: {rop.path()}",
-                detail=f"{exc}"
-            )
